Butchering stuff.
